0.0.7

+++ src/main/java/cn/mcmod/sakura/block/crop/BlockCabbageCrop.java
@@ -22,7 +22,7 @@ public class BlockCabbageCrop extends BlockCrops {
 	@Override
     public AxisAlignedBB getBoundingBox(IBlockState state, IBlockAccess source, BlockPos pos)
     {
-        return CABBAGE_AABB[((Integer)state.getValue(this.getAgeProperty())).intValue()];
+        return CABBAGE_AABB[state.getValue(this.getAgeProperty()).intValue()];
     }
 	@Override
 	protected Item getCrop() {
